[IE TESTS] Replaced size_t with int64_t (#2699)
* Replaced size_t with int64_t

* Fix gpu plugin

+++ inference-engine/tests/functional/plugin/shared/include/single_layer_tests/tile.hpp
@@ -14,7 +14,7 @@
 
 namespace LayerTestsDefinitions {
 
-typedef std::vector<size_t> TileSpecificParams;
+typedef std::vector<int64_t> TileSpecificParams;
 typedef std::tuple<
         TileSpecificParams,
         InferenceEngine::Precision,    // Net precision
