+++ inference-engine/tests/ngraph_functions/src/tile.cpp
@@ -8,7 +8,7 @@ namespace ngraph {
 namespace builder {
 
 std::shared_ptr<ngraph::Node> makeTile(const ngraph::Output<Node>& in,
-                                       const std::vector<size_t>& repeats) {
+                                       const std::vector<int64_t>& repeats) {
     auto repeatsNode = std::make_shared<ngraph::opset1::Constant>(ngraph::element::i64, std::vector<size_t>{repeats.size()}, repeats);
     auto tileNode = std::make_shared<ngraph::opset1::Tile>(in, repeatsNode);
     return tileNode;
